+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.2] - 2021-??-?? :question:
+* Fixes wrong mime type in OpenAPI Documentation for the `form` binder (#212)
+* Adds `OpenAPIEvents` to the class handling the generation of OpenAPI Documentation
+
 ## [1.2.1] - 2021-11-15 :shield:
 - Adds built-in support for `JWT` bearer authentication, and validation
   of `JWTs` issued by identity providers implementing **OpenID Connect (OIDC)**